+++ src/main/java/me/pajic/simple_smithing_overhaul/mixson/AssetPatches.java
@@ -2,6 +2,7 @@ package me.pajic.simple_smithing_overhaul.mixson;
 
 import me.pajic.simple_smithing_overhaul.SSO;
 import net.minecraft.client.Minecraft;
+import net.ramixin.mixson.util.Index;
 
 public class AssetPatches {
 
@@ -15,6 +16,15 @@ public class AssetPatches {
 							context.getFile().getAsJsonObject().remove("entity.minecraft.experience_bottle");
 							context.getFile().getAsJsonObject().remove("item.minecraft.experience_bottle");
 						}
+					}
+				}
+		);
+		MixsonHelper.registerSingleJson(
+				"Modify Mending description",
+				new Index("item_descriptions:lang/en_us"),
+				context -> {
+					if (SSO.CONFIG.mendingRework.enabled.get()) {
+						context.getFile().getAsJsonObject().remove("enchantment.minecraft.mending.description");
 					}
 				}
 		);